(Finally) Finish Channels
