+++ TwitchLib/TwitchAPI/Channels.cs
@@ -2,7 +2,6 @@
 using System.Linq;
 using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
-using TwitchLib.Exceptions;
 
 namespace TwitchLib.TwitchAPI
 {
@@ -52,37 +51,27 @@ namespace TwitchLib.TwitchAPI
         }
 
         /// <summary>
-        ///     Retrieves an ascending or descending list of followers from a specific channel.
+        ///     Returns a list of user objects who are editors of <paramref name="channel" />.
+        ///     <para>Authenticated, required scope: <code>channel_read</code></para>
         /// </summary>
-        /// <param name="channel">The channel to retrieve the followers from.</param>
-        /// <param name="limit">Maximum number of objects in array. Default is 25. Maximum is 100.</param>
-        /// <param name="cursor">
-        ///     Twitch uses cursoring to paginate long lists of followers. Check <code>_cursor</code> in response
-        ///     body and set <code>cursor</code> to this value to get the next page of results, or use <code>_links.next</code> to
-        ///     navigate to the next page of results.
-        /// </param>
-        /// <param name="direction">Creation date sorting direction.</param>
-        /// <returns>A list of TwitchFollower objects.</returns>
-        public static async Task<List<TwitchFollower>> GetTwitchFollowers(string channel, int limit = 25,
-            int cursor = -1, SortDirection direction = SortDirection.Descending)
+        /// <param name="channel">The channel to retrieve the editors from.</param>
+        /// <param name="accessToken">An oauth token with the required scope.</param>
+        /// <returns>A list of TwitchUser objects whom are channel editors.</returns>
+        public static async Task<List<TwitchUser>> GetChannelEditors(string channel, string accessToken)
         {
-            var args = "";
+            var resp = await MakeGetRequest($"{KrakenBaseUrl}/channels/{channel}/editors", accessToken);
+            var editors = JObject.Parse(resp).SelectToken("users");
 
-            args += "?limit=" + limit;
-            args += cursor != -1 ? $"&cursor={cursor}" : "";
-            args += "&direction=" + (direction == SortDirection.Descending ? "desc" : "asc");
-
-            var resp = await MakeGetRequest($"{KrakenBaseUrl}/channels/{channel}/follows{args}");
-            return JObject.Parse(resp).SelectToken("follows").Select(follower => new TwitchFollower(follower)).ToList();
+            return editors.Select(editor => new TwitchUser(editor)).ToList();
         }
 
         /// <summary>
         ///     Updates the <paramref name="delay" /> of a <paramref name="channel" />.
         ///     <para>Authenticated, required scope: <code>channel_editor</code></para>
         /// </summary>
-        /// <param name="delay">Channel delay in seconds.</param>
+        /// <param name="delay">Channel delay in seconds. Requires the channel owner's OAuth token.</param>
         /// <param name="channel">The channel to update.</param>
-        /// <param name="accessToken">The channel owner's access token and the required scope.</param>
+        /// <param name="accessToken">An oauth token with the required scope.</param>
         /// <returns>The response of the request.</returns>
         public static async Task<string> UpdateStreamDelay(int delay, string channel, string accessToken)
         {
@@ -98,7 +87,7 @@ namespace TwitchLib.TwitchAPI
         /// <param name="channel">The channel to update.</param>
         /// <param name="accessToken">An oauth token with the required scope.</param>
         /// <returns>The response of the request.</returns>
-        public static async Task<string> UpdateStreamTitle(string status, string channel, string accessToken)
+        public static async Task<string> UpdateStreamStatus(string status, string channel, string accessToken)
         {
             var data = "{\"channel\":{\"status\":\"" + status + "\"}}";
             return await MakeRestRequest($"{KrakenBaseUrl}/channels/{channel}", "PUT", data, accessToken);
@@ -126,7 +115,7 @@ namespace TwitchLib.TwitchAPI
         /// <param name="channel">The channel to update.</param>
         /// <param name="accessToken">An oauth token with the required scope.</param>
         /// <returns>The response of the request.</returns>
-        public static async Task<string> UpdateStreamTitleAndGame(string status, string game, string channel,
+        public static async Task<string> UpdateStreamStatusAndGame(string status, string game, string channel,
             string accessToken)
         {
             var data = "{\"channel\":{\"status\":\"" + status + "\",\"game\":\"" + game + "\"}}";
@@ -144,28 +133,6 @@ namespace TwitchLib.TwitchAPI
         {
             return await
                 MakeRestRequest($"{KrakenBaseUrl}/channels/{channel}/streamkey", "DELETE", "", accessToken);
-        }
-
-        /// <summary>
-        ///     Returns a list of videos ordered by time of creation, starting with the most recent.
-        /// </summary>
-        /// <param name="channel">The channel to retrieve the list of videos from.</param>
-        /// <param name="limit">Maximum number of objects in array. Default is 10. Maximum is 100.</param>
-        /// <param name="offset">Object offset for pagination. Default is 0.</param>
-        /// <param name="onlyBroadcasts">
-        ///     Returns only broadcasts when true. Otherwise only highlights are returned. Default is
-        ///     false.
-        /// </param>
-        /// <param name="onlyHls">Returns only HLS VoDs when true. Otherwise only non-HLS VoDs are returned. Default is false.</param>
-        /// <returns>A list of TwitchVideo objects the channel has available.</returns>
-        public static async Task<List<TwitchVideo>> GetChannelVideos(string channel, int limit = 10,
-            int offset = 0, bool onlyBroadcasts = false, bool onlyHls = false)
-        {
-            var args = $"?limit={limit}&offset={offset}&broadcasts={onlyBroadcasts}&hls={onlyHls}";
-            var resp = await MakeGetRequest($"{KrakenBaseUrl}/channels/{channel}/videos{args}");
-            var vids = JObject.Parse(resp).SelectToken("videos");
-
-            return vids.Select(vid => new TwitchVideo(vid)).ToList();
         }
 
         /// <summary>
